cambios para interfaz, después sigo que ya está el café.

+++ Proyecto Dots&Boxes P2/Proyecto Dots&Boxes P2/stratCercano.cpp
@@ -1,7 +1,7 @@
 #include "stratCercano.h"
 
 bool stratCercano::strat(tabla* mat, vecPosicionesLibres& vec) {
-	srand(time(0));
+	srand((long)time(0));
 	int x = rand() % (vec.getCan() + 1);
 
 	if (vec.getPosLibre(x) != nullptr) {
